+++ processor/src/test/resources/expected/com/example/repository/Arez_DaggerEnabledRepository.java
@@ -15,10 +15,10 @@ import javax.annotation.Nullable;
 import org.realityforge.braincheck.Guards;
 
 @Generated("arez.processor.ArezProcessor")
-final class Arez_DaggerEnabledRepository extends DaggerEnabledRepository implements Disposable, Identifiable<Long>, ComponentObservable {
-  private static volatile long $$arezi$$_nextId;
+final class Arez_DaggerEnabledRepository extends DaggerEnabledRepository implements Disposable, Identifiable<Integer>, ComponentObservable {
+  private static volatile int $$arezi$$_nextId;
 
-  private final long $$arezi$$_id;
+  private final int $$arezi$$_id;
 
   private byte $$arezi$$_state;
 
@@ -54,13 +54,13 @@ final class Arez_DaggerEnabledRepository extends DaggerEnabledRepository impleme
     return Arez.areZonesEnabled() ? this.$$arezi$$_context : Arez.context();
   }
 
-  final long $$arezi$$_id() {
+  final int $$arezi$$_id() {
     return this.$$arezi$$_id;
   }
 
   @Override
   @Nonnull
-  public final Long getArezId() {
+  public final Integer getArezId() {
     return $$arezi$$_id();
   }
 
@@ -130,7 +130,7 @@ final class Arez_DaggerEnabledRepository extends DaggerEnabledRepository impleme
 
   @Override
   public final int hashCode() {
-    return Long.hashCode( $$arezi$$_id() );
+    return Double.hashCode( $$arezi$$_id() );
   }
 
   @Override
